requirements.txt
asgiref==3.2.7
Babel==2.8.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
decorator==4.4.2
distro==1.5.0
Django==3.0.4
django-dynamic-preferences==1.8.1
django-environ==0.4.5
django-mailgun==0.9.1
django-nested-admin==3.2.4
django-phonenumber-field==4.0.0
django-redis==4.11.0
djangorestframework==3.11.0
future==0.18.2
humanize==2.3.0
idna==2.9
loguru==0.4.1
persisting-theory==0.2.1
phonenumbers==8.12.0
Pillow==7.1.1
psycopg2-binary==2.8.4
pycparser==2.20
python-datauri==0.2.8
python-dateutil==2.8.1
python-monkey-business==1.0.0
python-telegram-bot==12.5.1
pytz==2019.3
PyYAML==5.3.1
redis==3.4.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
tornado==6.0.4
urllib3==1.25.8
uuid==1.30
uWSGI==2.0.18
yandex-checkout==1.5.1
